embeded/requirements.txt
awscrt==0.21.1
awsiotsdk==1.22.0
boto3==1.35.54
botocore==1.35.54
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-vision==3.8.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
jmespath==1.0.1
numpy==1.24.4
opencv-python==4.10.0.84
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
six==1.16.0
urllib3==1.26.20
